refactor(database): normalize initial migration schema conventions
Apply the consistency batch from the migration cleanup plan to
01_initial.up.sql:

- convert the three SERIAL outbox PKs (druckauftraege,
  tse_nachsignier_auftraege, tse_signaturen) to INT GENERATED BY
  DEFAULT AS IDENTITY
- normalize INTEGER -> INT and VARCHAR(50) -> TEXT
- drop OR REPLACE from all eleven function definitions
- delete redundant indexes idx_users_username (username is UNIQUE) and
  idx_tisch_favoriten_user_id (PK already leads with user_id)
- rename the abbreviated idx_*_ks_nr indexes to idx_*_kassensitzung_nr
- fix the admin INSERT comment (onetime password IS set) and stray
  whitespace
- add COMMENT ON COLUMN for every tisch_sessions and tse_signaturen
  column

The SERIAL->IDENTITY conversion makes the three outbox id columns hit
the pg_catalog.int4 -> int override in sqlc.yaml (SERIAL was catalogued
as serial4 and fell back to int32), so regenerated dbgen code now types
them int like every other ID; the now-redundant int32() bridge casts in
druckauftrag_repo and tse_repo are removed.

Phase 2 of docs/plans/plan-migrations-cleanup.md (completes the plan).

+++ backend/repository/druckauftrag_repo/repo.go
@@ -104,7 +104,7 @@ func (r Repository) GetOffeneDruckauftraege(ctx context.Context) ([]OffenerDruck
 	result := make([]OffenerDruckauftrag, 0, len(rows))
 	for _, row := range rows {
 		result = append(result, OffenerDruckauftrag{
-			ID:      int(row.ID),
+			ID:      row.ID,
 			ZielIP:  row.ZielIp,
 			Payload: row.Payload,
 		})
@@ -131,13 +131,13 @@ func (r Repository) MeldeDruckergebnis(ctx context.Context, gedruckteIDs []int, 
 
 	qtx := r.q.WithTx(tx)
 	for _, id := range gedruckteIDs {
-		if err := qtx.MarkDruckauftragGedruckt(ctx, int32(id)); err != nil {
+		if err := qtx.MarkDruckauftragGedruckt(ctx, id); err != nil {
 			return db.Error(err)
 		}
 	}
 	for _, f := range fehlversuche {
 		err := qtx.IncrementDruckauftragFehlversuch(ctx, dbgen.IncrementDruckauftragFehlversuchParams{
-			ID:            int32(f.ID),
+			ID:            f.ID,
 			LetzterFehler: sql.NullString{String: f.Fehler, Valid: true},
 			MaxVersuche:   MaxDruckversuche,
 		})
@@ -164,7 +164,7 @@ func (r Repository) GetFehlgeschlageneDruckauftraege(ctx context.Context) ([]Feh
 	result := make([]FehlgeschlagenerDruckauftrag, 0, len(rows))
 	for _, row := range rows {
 		result = append(result, FehlgeschlagenerDruckauftrag{
-			ID:            int(row.ID),
+			ID:            row.ID,
 			BonArt:        row.BonArt,
 			ZielIP:        row.ZielIp,
 			Referenz:      row.Referenz,
@@ -181,12 +181,12 @@ func (r Repository) GetFehlgeschlageneDruckauftraege(ctx context.Context) ([]Feh
 // (fehlgeschlagen -> offen, versuche zurück auf 0). Der Status-Guard wirkt nur
 // auf fehlgeschlagene Aufträge; andere Status bleiben unberührt.
 func (r Repository) DruckauftragErneutVersuchen(ctx context.Context, id int) error {
-	return db.Error(r.q.DruckauftragErneutVersuchen(ctx, int32(id)))
+	return db.Error(r.q.DruckauftragErneutVersuchen(ctx, id))
 }
 
 // DruckauftragVerwerfen markiert einen fehlgeschlagenen Auftrag als verworfen
 // (fehlgeschlagen -> verworfen). Der Eintrag bleibt in der Datenbank erhalten;
 // der Status-Guard wirkt nur auf fehlgeschlagene Aufträge.
 func (r Repository) DruckauftragVerwerfen(ctx context.Context, id int) error {
-	return db.Error(r.q.DruckauftragVerwerfen(ctx, int32(id)))
+	return db.Error(r.q.DruckauftragVerwerfen(ctx, id))
 }
